+++ internal/builder/scripts/initconf.sh
@@ -8,21 +8,7 @@ SLURM_USER="slurm"
 SLURM_MOUNT=/mnt/slurm
 SLURM_DIR=/mnt/etc/slurm
 
-function init::common() {
-	local dir
-
-	dir=/var/spool/slurmd
-	mkdir -p "$dir"
-	chown -v "${SLURM_USER}:${SLURM_USER}" "$dir"
-	chmod -v 700 "$dir"
-
-	dir=/var/spool/slurmctld
-	mkdir -p "$dir"
-	chown -v "${SLURM_USER}:${SLURM_USER}" "$dir"
-	chmod -v 700 "$dir"
-}
-
-function init::slurm() {
+function main() {
 	# Workaround to ephemeral volumes not supporting securityContext
 	# https://github.com/kubernetes/kubernetes/issues/81089
 
@@ -30,15 +16,11 @@ function init::slurm() {
 	mkdir -p "$SLURM_DIR"
 	find "${SLURM_MOUNT}" -type f -name "*.conf" -print0 | xargs -0r cp -vt "${SLURM_DIR}"
 	find "${SLURM_MOUNT}" -type f -name "*.key" -print0 | xargs -0r cp -vt "${SLURM_DIR}"
-	find "${SLURM_MOUNT}" -type f -regextype posix-extended -regex "^.*/(pro|epi)log-.*$" -print0 | xargs -0r cp -vt "${SLURM_DIR}"
-	find "${SLURM_MOUNT}" -type f -regextype posix-extended -regex "^.*/(pro|epi)log-.*$" -print0 | xargs -0r cp -vt "${SLURM_DIR}"
 
 	# Set general permissions and ownership
 	find "${SLURM_DIR}" -type f -print0 | xargs -0r chown -v "${SLURM_USER}:${SLURM_USER}"
 	find "${SLURM_DIR}" -type f -name "*.conf" -print0 | xargs -0r chmod -v 644
 	find "${SLURM_DIR}" -type f -name "*.key" -print0 | xargs -0r chmod -v 600
-	find "${SLURM_DIR}" -type f -regextype posix-extended -regex "^.*/(pro|epi)log-.*$" -print0 | xargs -0r chown -v "${SLURM_USER}:${SLURM_USER}"
-	find "${SLURM_DIR}" -type f -regextype posix-extended -regex "^.*/(pro|epi)log-.*$" -print0 | xargs -0r chmod -v 755
 
 	# Inject secrets into certain config files
 	local dbd_conf="slurmdbd.conf"
@@ -52,10 +34,5 @@ function init::slurm() {
 
 	# Display Slurm directory files
 	ls -lAF "${SLURM_DIR}"
-}
-
-function main() {
-	init::common
-	init::slurm
 }
 main